libmems: Fix limit ODR
Fix initialization error that crash libmems in the iio_context
constructor.
Fix type.

BUG=b:353354908,b:406277521
TEST=Unit test, Check `cros-health-tool diag sensitive_sensor_v2` on
dewatt.

Fixes: 0f9d0fd0a74 ("libmems: limit ODR")
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6405837
Auto-Submit: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Gwendal Grignou <gwendal@chromium.org>

+++ libmems/iio_device.cc
@@ -218,7 +218,7 @@ bool IioDevice::GetMinMaxFrequency(double* min_freq, double* max_freq) {
         }
       }
   }
-  std::optional<int> max_sensor_odr = GetContext()->GetMaxSensorOdr();
+  std::optional<double> max_sensor_odr = GetContext()->GetMaxSensorOdr();
   if (max_sensor_odr && *max_sensor_odr < *max_freq) {
     *max_freq = *max_sensor_odr;
   }
